Moved list related transformers from projectTransformer to transformers #565
